WIP list_reverse

+++ linkedlist.c
@@ -198,7 +198,7 @@ void list_reverse(list_t *list) {
 	node_t *beg = list->head, *end = list->tail;
 	void *tmp;
 	for (int i = 0; i < list_size(list) / 2; i++) {
-		intptr_t *a = beg->item, *b = end->item;
+		int *a = beg->item, *b = end->item;
 		*a ^= *b ^= *a ^= *b;
 		beg->item = a, end->item = b;
 		// (int *)beg->item ^= (int *)end->item;
